+++ ludwig/backend/ray.py
@@ -148,10 +148,12 @@ _engine_registry = {
 
 
 def _get_df_engine(processor):
+    logger.info(f"Ray processor params: {processor}")
     if processor is None:
+        # TODO ray: find an informed way to set the parallelism, in practice
+        #  it looks like Dask handles this well on its own most of the time
         return _create_dask_engine()
 
-    logger.info(f"Ray processor params: {processor}")
     processor_kwargs = processor.copy()
 
     dtype = processor_kwargs.pop("type", "dask")
