+++ ludwig/hyperopt/run.py
@@ -2,7 +2,7 @@ import copy
 import logging
 import os
 from pprint import pformat
-from typing import Any, Dict, List, Optional, Union
+from typing import List, Optional, Union
 
 import pandas as pd
 import yaml
@@ -12,22 +12,15 @@ from ludwig.api import LudwigModel
 from ludwig.backend import Backend, initialize_backend, LocalBackend
 from ludwig.callbacks import Callback
 from ludwig.constants import (
-    AUTO,
     COMBINED,
     EXECUTOR,
-    GOAL,
     HYPEROPT,
     LOSS,
-    MAX_CONCURRENT_TRIALS,
-    METRIC,
-    MINIMIZE,
     NAME,
     NUM_SAMPLES,
     OUTPUT_FEATURES,
     PARAMETERS,
     PREPROCESSING,
-    RAY,
-    SPLIT,
     TEST,
     TRAINING,
     TYPE,
@@ -41,13 +34,14 @@ from ludwig.hyperopt.utils import (
     print_hyperopt_results,
     save_hyperopt_stats,
     should_tune_preprocessing,
+    update_hyperopt_params_with_defaults,
     update_or_set_max_concurrent_trials,
 )
 from ludwig.utils.backward_compatibility import upgrade_to_latest_version
 from ludwig.utils.dataset_utils import generate_dataset_statistics
 from ludwig.utils.defaults import default_random_seed, merge_with_defaults
 from ludwig.utils.fs_utils import makedirs, open_file
-from ludwig.utils.misc_utils import get_class_attributes, get_from_registry, set_default_value, set_default_values
+from ludwig.utils.misc_utils import get_from_registry
 
 try:
     from ludwig.backend.ray import RayBackend
@@ -410,22 +404,3 @@ def hyperopt(
     logger.info("Finished hyperopt")
 
     return hyperopt_results
-
-
-def update_hyperopt_params_with_defaults(hyperopt_params: Dict[str, Any]) -> None:
-    from ludwig.hyperopt.execution import executor_registry
-
-    set_default_value(hyperopt_params, EXECUTOR, {})
-    set_default_value(hyperopt_params, SPLIT, VALIDATION)
-    set_default_value(hyperopt_params, "output_feature", COMBINED)
-    set_default_value(hyperopt_params, METRIC, LOSS)
-    set_default_value(hyperopt_params, GOAL, MINIMIZE)
-
-    set_default_values(hyperopt_params[EXECUTOR], {TYPE: RAY, NUM_SAMPLES: 1, MAX_CONCURRENT_TRIALS: AUTO})
-
-    executor = get_from_registry(hyperopt_params[EXECUTOR][TYPE], executor_registry)
-    executor_defaults = {k: v for k, v in executor.__dict__.items() if k in get_class_attributes(executor)}
-    set_default_values(
-        hyperopt_params[EXECUTOR],
-        executor_defaults,
-    )
